backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-auth-oauthlib==1.1.0
google-auth==2.23.4
google-api-python-client==2.108.0
python-multipart==0.0.6

# Database
sqlmodel
pydantic-settings
sqlalchemy>=2.0.36

# Image processing
pillow
pillow-heif

# Additional utilities
requests==2.31.0

# Phase 1 — multi-user platform
# Token encryption for DriveConnection.encrypted_*_token fields
# Generate a key: python -c "from cryptography.fernet import Fernet; print(Fernet.generate_key().decode())"
cryptography>=42.0.0

# AI (optional — only needed when ai_enabled=true)
# openai>=1.0.0
# langchain-openai 